feat: modularize data stream form view
